Move pom.xml spec to a specific go file
